fix: heading levels
The heading levels are really wierd.

ExpressLRS is a second H1 and I think it intended to be a sub headers

Yet also it's the only subheader and all other scripts are beneath it.

This commit changes the README so ExpressLRS is a H2 and adds others (somewhat correctly) while moving the links to H3.